change dump test to include tempfile
