+++ tests/test_util.py
@@ -3,9 +3,15 @@ import urllib
 from pathlib import Path
 
 import pytest
+from ruamel.yaml import YAML
 
 from ib_manifest_util import TEMPLATE_DIR, TEST_DATA_DIR
-from ib_manifest_util.util import download_files, write_templatized_file
+from ib_manifest_util.util import (
+    download_files,
+    dump_yaml,
+    load_yaml,
+    write_templatized_file,
+)
 
 
 def test_write_templatized_file_hardening():
@@ -117,3 +123,78 @@ def test_download_file_multiple_urls(cleanup):
         f_path = Path(f_name).resolve()
         cleanup.append(f_path)
         assert f_path.exists(), f"File should be written to {f_path}."
+
+
+def load_yaml_for_testing(file_path: str | Path) -> dict:
+    """Load a yaml file.
+
+    This provides a method for loading yaml files independent of utils.load_yaml.
+    Args:
+        file_path: str | Path
+            Path to yaml file.
+    Returns: dict
+    """
+
+    yaml_loader = YAML(typ="safe")
+    file_path = Path(file_path).resolve()
+    with open(file_path, "r") as f:
+        return yaml_loader.load(f)
+
+
+def test_dump_yaml(cleanup):
+    """Test that dictionary is correctly written to .yaml file"""
+    sample_yaml_path = TEST_DATA_DIR.joinpath("sample_yaml.yaml")
+    tempfile_path = Path("tempfile.yaml").resolve()
+    cleanup.append(tempfile_path)
+
+    sample_yaml_dict = load_yaml_for_testing(sample_yaml_path)
+
+    # Test path type Path
+    dump_yaml(sample_yaml_dict, target_path=tempfile_path)
+    tempfile_dict = load_yaml_for_testing(tempfile_path)
+    assert (
+        tempfile_dict["myKey"] == "myValue"
+    ), "Loaded yaml file should provide correct key-value pair."
+
+
+def test_load_yaml_path_types():
+    """Test load_yaml file_path arg"""
+    hardening_manifest_path = TEST_DATA_DIR.joinpath("hardening_manifest.yaml")
+
+    # Test path type str
+    hardening_manifest = load_yaml(file_path=str(hardening_manifest_path))
+    assert (
+        hardening_manifest["apiVersion"] == "v1"
+    ), "Loaded yaml file should provide correct key-value pair."
+
+    # Test path type Path
+    hardening_manifest = load_yaml(file_path=hardening_manifest_path)
+    assert (
+        hardening_manifest["apiVersion"] == "v1"
+    ), "Loaded yaml file should provide correct key-value pair."
+
+    # Test missing file
+    with pytest.raises(FileNotFoundError):
+        load_yaml("this_file_does_not_exist.yaml")
+
+
+def test_load_yaml_return_type():
+    """Test load_yaml return type"""
+    hardening_manifest_path = TEST_DATA_DIR.joinpath("hardening_manifest.yaml")
+    hardening_manifest = load_yaml(file_path=hardening_manifest_path)
+    assert isinstance(
+        hardening_manifest, dict
+    ), "Returned object should be a dictionary."
+
+
+def test_load_yaml_loaders():
+    """Test load_yaml loader_type arg"""
+    hardening_manifest_path = TEST_DATA_DIR.joinpath("hardening_manifest.yaml")
+
+    # Test safe load
+    hardening_manifest = load_yaml(
+        file_path=hardening_manifest_path, loader_type="safe"
+    )
+    assert (
+        hardening_manifest["apiVersion"] == "v1"
+    ), "Loaded yaml file should provide correct key-value pair."
